requirements.txt
pandas==2.2.3
geopandas==1.0.1
numpy==2.2.6
requests==2.32.3
python-dotenv==1.1.0
openpyxl==3.1.5